test: add windows UT support and multiple tests and VS projects
Also misc modifications to support windows compilation. Not all
tests are included in this patch, just the required framework,
two sets of tests, blk_nblock and blk_non_zero, and several but
not all VS project files for the tests.

+++ src/common/set.c
@@ -58,7 +58,7 @@
 #include "out.h"
 #include "valgrind_internal.h"
 
-extern unsigned long Pagesize;
+extern unsigned long long Pagesize;
 
 /* reserve space for size, path and some whitespace and/or comment */
 #define PARSER_MAX_LINE (PATH_MAX + 1024)
